requirements.txt
tensorboardX
torch
numpy
open3d==0.16.0
pymcubes
plyfile
h5py
transforms3d
tqdm
opencv-python
pyyaml
scikit-image
trimesh
xatlas